Common interface

MolecularGraph.Graph.addnode! — Method

addnode!(graph) -> Int
addnode!(graph, attr) -> Int

Add new node and return the node index. If the node attribute type is required, specify the node attribute object by node keyword.

MolecularGraph.Graph.clearcache! — Method

clearcache!(graph::AbstractGraph)

Clear calculated property caches.

Calling clearcache! is recommended when the graph nodes/edges are added, removed or reindexed. You can deepcopy or Graph.clone the graph before destructive operation instead.

MolecularGraph.Graph.clone — Function

clone(graph::AbstractGraph) -> AbstractGraph

Return deep copy of the graph.

MolecularGraph.Graph.edgeattr — Method

edgeattr(graph::AbstractGraph, u, v) -> Union{AbstractEdge,Nothing}

Return the attribute object of an edge that connects u and v. If not found, return nothing.

MolecularGraph.Graph.edgeattr — Method

edgeattr(graph::AbstractGraph, i) -> AbstractEdge

Return the attribute object of edge i.

MolecularGraph.Graph.edgeattrs — Method

edgeattrs(graph::Union{OrderedGraph,OrderedDiGraph,OrderedHyperGraph}) -> Vector{AbstractEdge}

Return graph edge attributes.

MolecularGraph.Graph.edgeattrtype — Method

edgeattrtype(graph) -> Type

Return edge attribute type.

MolecularGraph.Graph.edgecount — Method

edgecount(graph::AbstractGraph) -> Int

Return the number of graph edges.

MolecularGraph.Graph.edgeset — Method

edgeset(graph::Union{OrderedGraph,OrderedDiGraph,OrderedHyperGraph}) -> Set{Int}

Return the set of edge keys.

MolecularGraph.Graph.findalledgekeys — Method

findalledgekeys(graph::UndirectedGraph, u, v) -> Any
findalledgekeys(graph::DirectedGraph, u, v) -> Any
findalledgekeys(graph::HyperGraph, u, v) -> Any

Return all edges that connect nodes u and v.

MolecularGraph.Graph.findedgekey — Method

findedgekey(graph::UndirectedGraph, u, v) -> Any
findedgekey(graph::DirectedGraph, u, v) -> Any
findedgekey(graph::HyperGraph, u, v) -> Any

Return the first edge found that connects nodes u and v.

MolecularGraph.Graph.getcache — Method

getcache(graph::AbstractGraph, key) -> Any

Get cached calculated properties.

MolecularGraph.Graph.getedge — Method

getedge(graph::AbstractGraph, i) -> Any

Return an edge.

MolecularGraph.Graph.hascache — Method

hascache(graph::AbstractGraph, key) -> Bool

Return whether the graph has calculated properties with the given key or not.

MolecularGraph.Graph.hascachestore — Method

hascachestore(graph::AbstractGraph) -> Bool

Return whether the graph has calculated property cache store.

MolecularGraph.Graph.hasedge — Method

hasedge(graph, u, v) -> Bool

Return whether the given two nodes are connected by at least one edge.

MolecularGraph.Graph.indegree — Method

indegree(graph::DirectedGraph, n) -> Int

Return the number of inneighbors of the node 'n'.

MolecularGraph.Graph.inneighbors — Method

inneighbors(graph::DirectedGraph, i) -> Dict

Return the mapping of predecessor node keys and in edge keys connected to the given node.

MolecularGraph.Graph.neighborcount — Method

neighborcount(graph::AbstractGraph, n) -> Int
degree(graph::AbstractGraph, n) -> Int

Return the number of adjacent nodes of the node 'n'.

MolecularGraph.Graph.neighbors — Method

neighbors(graph, i) -> Dict

Return the mapping of incident edges and adjacent nodes of node i. If the graph is directed graph, both outneighbors and inneighbors are mapped.

MolecularGraph.Graph.nodeattr — Method

nodeattr(graph::AbstractGraph, i) -> AbstractNode

Return the attribute object of node i.

MolecularGraph.Graph.nodeattrs — Method

nodeattrs(graph::Union{OrderedGraph,OrderedDiGraph,OrderedHyperGraph}) -> Vector{AbstractNode}

Return graph node attributes.

MolecularGraph.Graph.nodeattrtype — Method

nodeattrtype(graph) -> Type

Return node attribute type.

MolecularGraph.Graph.nodecount — Method

nodecount(graph::AbstractGraph) -> Int

Return the number of graph nodes.

MolecularGraph.Graph.nodeset — Method

nodeset(graph::Union{OrderedGraph,OrderedDiGraph,OrderedHyperGraph}) -> Set{Int}

Return the set of node keys.

MolecularGraph.Graph.outdegree — Method

outdegree(graph::DirectedGraph, n) -> Int

Return the number of outneighbors of the node 'n'.

MolecularGraph.Graph.outneighbors — Method

outneighbors(graph::DirectedGraph, i) -> Dict

Return the mapping of successor node keys and out edge keys connected to the given node.

MolecularGraph.Graph.setcache! — Method

setcache!(graph::AbstractGraph, key; kwargs...)

Set calculated property caches.

MolecularGraph.Graph.setedgeattr! — Method

setedgeattr!(graph::AbstractGraph, i, attr::AbstractNode)

Update the edge attribute.

MolecularGraph.Graph.setnodeattr! — Method

setnodeattr!(graph::AbstractGraph, i, attr::AbstractNode)

Update the node attribute.

MolecularGraph.Graph.unlinkedges — Function

unlinkedges(graph, edges)

Delete given edges from the graph.

MolecularGraph.Graph.unlinknodes — Function

unlinknodes(graph, nodes)

Delete given nodes and its incident edges from the graph.

Graph model

Plain graph

MolecularGraph.Graph.immutableplaingraph — Method

immutableplaingraph() -> ImmutablePlainGraph
immutableplaingraph(size::Int, edges) -> ImmutablePlainGraph
immutableplaingraph(graph::UndirectedGraph) -> ImmutablePlainGraph

Convert an arbitrary UndirectedGraph into a ImmutablePlainGraph.

ImmutablePlainGraph is same as PlainGraph except for availability of addnode! and addedge!, and no performance difference exists between these graphs. ImmutablePlainGraph is mainly used for unit testing to check adverse effect of some graph algorithm implementations.

MolecularGraph.Graph.plaingraph — Method

plaingraph(nodes, edges) -> PlainGraph{PlainNode,Edge}

Generate vector graph that have given nodes and edges represented by the list of node indices in integer and the list of pairs of node indices, respectively.

MolecularGraph.Graph.plaingraph — Method

plaingraph(graph::OrderedGraph) -> PlainGraph

Convert an arbitrary OrderedGraph into a PlainGraph.

MolecularGraph.Graph.plaingraph — Method

plaingraph(graph::UndirectedGraph) -> PlainGraph

Convert an arbitrary UndirectedGraph into a PlainGraph.

If the given graph has non-sequential indices (ex. subgraph view), node and edge indices are sorted in ascending order and are re-indexed.

MolecularGraph.Graph.plaingraph — Method

plaingraph() -> PlainGraph

Generate empty PlainGraph.

Hyper graph

MolecularGraph.Graph.plainhypergraph — Method

plainhypergraph(size, edges) -> PlainHyperGraph

Generate hyper graph that have given size (number of nodes) and edges (Sets of nodes).

MolecularGraph.Graph.plainhypergraph — Method

plainhypergraph(graph::HyperGraph) -> PlainHyperGraph

Convert an arbitrary HyperGraph into a PlainHyperGraph.

If the given graph has non-sequential indices (ex. subgraph view), node and edge indices are sorted in ascending order and are re-indexed.

MolecularGraph.Graph.plainhypergraph — Method

plainhypergraph(graph::OrderedHyperGraph) -> PlainHyperGraph

Convert an arbitrary OrderedHyperGraph into a PlainHyperGraph.

MolecularGraph.Graph.plainhypergraph — Method

plainhypergraph() -> PlainHyperGraph

Generate empty PlainGraph.

Directed Acyclic Graph (DAG)

MolecularGraph.Graph.topologicalsort — Method

topologicalsort(graph::DirectedGraph) -> Vector{Int}
topologicalsort(graph::DirectedGraph, root) -> Vector{Int}

Run topological sort to obtain an array of sorted nodes.

If root is given, the method is applied to only the downstream nodes of root.

MolecularGraph.Graph.transitive_reduction — Method

transitive_reduction(graph::DirectedGraph) -> Set{Int}

Return transitive reduction as a subset of the DAG edges.

Graph generation

Graph generator

MolecularGraph.Graph.circularladder — Method

circularladder(n::Int; mutable=false) -> PlainGraph

Generate circular ladder graph $CL_n$.

MolecularGraph.Graph.completebipartite — Method

completebipartite(m::Int,n::Int; mutable=false) -> PlainGraph

Generate bipartite graph $K_{m,n}$.

MolecularGraph.Graph.completegraph — Method

completegraph(length::Int; mutable=false) -> PlainGraph

Generate complete graph $K_n$.

MolecularGraph.Graph.cyclegraph — Method

cyclegraph(length::Int; mutable=false) -> PlainGraph

Generate cycle graph $C_n$.

MolecularGraph.Graph.generalizedpetersen — Method

generalizedpetersen(n::Int, k::Int; mutable=false) -> PlainGraph

Generate generalized petersen graph $G(n,k)$.

MolecularGraph.Graph.laddergraph — Method

laddergraph(n::Int; mutable=false) -> PlainGraph

Generate ladder graph $L_n$.

MolecularGraph.Graph.moebiusladder — Method

moebiusladder(n::Int; mutable=false) -> PlainGraph

Generate Möbius ladder graph $ML_n$.

MolecularGraph.Graph.pathgraph — Method

pathgraph(n::Int; mutable=false) -> PlainGraph

Generate path graph $P_n$.

MolecularGraph.Graph.squaregrid — Method

squaregrid(m::Int,n::Int; mutable=false) -> PlainGraph

Generate $m \times n$ square grid graph.

Use cartesianproduct for higher dimensional grid graphs.

Product of graphs

MolecularGraph.Graph.cartesianproduct — Method

cartesianproduct(G::OrderedGraph, H::OrderedGraph) -> CartesianProduct

Return the cartesian product of graphs G and H.

MolecularGraph.Graph.modularproduct — Method

modularproduct(G::OrderedGraph, H::OrderedGraph) -> ModularProduct

Return the modular product of graphs G and H.

Disjoint union

MolecularGraph.Graph.disjointunion! — Method

disjointunion!(g1::T, g2::T, G::T...) where {T<:OrderedGraph} -> T

Generate disjoint union graph of given graphs. g1 will be overwritten by the union graph. Unlike non-destructive disjointunion, g1 does not retain any information about other given graphs but a bit faster.

MolecularGraph.Graph.disjointunion — Method

disjointunion(g1::UndirectedGraph, g2::UndirectedGraph,
    G::UndirectedGraph...) -> DisjointUnionGraph

Generate disjoint union graph of given graphs. The new graph with type DisjointUnionGraph retains mapping to the original graphs as nodes and edges attributes.

Contraction

MolecularGraph.Graph.edgecontraction — Method

edgecontraction(graph::UndirectedGraph, nodes) -> ContractionView

Generate edge-contraction view.

Line graph

MolecularGraph.Graph.linegraph — Method

linegraph(G::AbstractGraph) -> LineGraph

Generate line graph.

Subgraph view

MolecularGraph.Graph.edgesubgraph — Method

edgesubgraph(graph::UndirectedGraph, edges) -> SubgraphView
edgesubgraph(graph::DirectedGraph, edges) -> DiSubgraphView

Generate edge-induced subgraph view.

MolecularGraph.Graph.nodesubgraph — Method

nodesubgraph(graph::UndirectedGraph, nodes) -> SubgraphView
nodesubgraph(graph::DirectedGraph, nodes) -> DiSubgraphView

Generate node-induced subgraph view.

MolecularGraph.Graph.subgraphview — Method

subgraphview(graph::UndirectedGraph, nodes, edges) -> SubgraphView
subgraphview(graph::DirectedGraph, nodes, edges) -> DiSubgraphView

Generate subgraph view with the given nodes and edges.

MolecularGraph.Graph.supergraph — Method

supergraph(view::SubgraphView) -> UndirectedGraph
supergraph(view::DiSubgraphView) -> DirectedGraph

Returns the graph on which the view is based.

Graph algorithms

Connectivity

MolecularGraph.Graph.bridges — Method

bridges(graph::UndirectedGraph) -> Set{Int}
bridges(state::BiconnectedState) -> Set{Int}

Compute biconnectivity and return bridges.

MolecularGraph.Graph.connectedcomponents — Method

connectedcomponents(graph::UndirectedGraph) -> Vector{Set{Int}}

Compute connectivity and return sets of the connected components.

MolecularGraph.Graph.connectedmembership — Method

connectedmembership(graph::UndirectedGraph) -> Vector{Int}

Return connected component membership array.

MolecularGraph.Graph.cutvertices — Method

cutvertices(graph::UndirectedGraph) -> Set{Int}
cutvertices(state::BiconnectedState) -> Set{Int}

Compute biconnectivity and return cut vertices (articulation points).

MolecularGraph.Graph.edgebiconnectedcomponents — Method

edgebiconnectedcomponents(graph::UndirectedGraph) -> Vector{Vector{Int}}

Compute sets of biconnected component edges.

Shortest path

MolecularGraph.Graph.diameter — Method

diameter(graph::AbstractGraph) -> Int

Compute the diameter of the graph (the largest eccentricity of any nodes).

MolecularGraph.Graph.distance — Method

distance(graph::AbstractGraph, source::Int, target::Int) -> Int
reversedistance(graph::DirectedGraph, source::Int, target::Int) -> Int

Compute the distance (shortest path length) from source to target. If the nodes are not reachable each other, the value will be nothing.

MolecularGraph.Graph.distancematrix — Method

distancematrix(graph::OrderedGraph) -> Matrix{Float64}

Generate the distance matrix of the graph.

Note that the type of the generated matrix will be Float64. If the nodes are not reachable each other, the distance value will be Inf.

MolecularGraph.Graph.eccentricity — Method

eccentricity(graph::UndirectedGraph, v::Int) -> Int

Compute the eccentricity of the graph (the largest distance between v and any other nodes).

MolecularGraph.Graph.isreachable — Method

reachablenodes(graph::AbstractGraph, u::Int, v::Int) -> Bool
reversereachablenodes(graph::DirectedGraph, u::Int, v::Int) -> Bool

Return whether the node v is reachable from u.

MolecularGraph.Graph.longestshortestpathnodes — Method

longestshortestpathnodes(graph::UndirectedGraph) -> Vector{Int}

Compute the longest shortest path in the graph (a path between two arbitrary peripheral nodes) as a vector of nodes that starts with one of the peripheral node and ends with the other side.

MolecularGraph.Graph.reachablenodes — Method

reachablenodes(graph::AbstractGraph, node::Int) -> Set{Int}
reversereachablenodes(graph::DirectedGraph, node::Int) -> Set{Int}

Return the set of reachable nodes from node.

MolecularGraph.Graph.shortestpathedges — Method

shortestpathedges(graph::UndirectedGraph, u::Int, v::Int) -> Vector{Int}

Compute the shortest path between u and v as a vector of the edges that forms the path. Return nothing if not reachable.

MolecularGraph.Graph.shortestpathnodes — Method

shortestpathnodes(graph::UndirectedGraph, u::Int, v::Int) -> Vector{Int}

Compute the shortest path between u and v as a vector of the nodes that forms the path. Return nothing if not reachable.

Cycle basis

MolecularGraph.Graph.minimumcyclebasis — Method

minimumcyclebasis(graph::UndirectedGraph) -> Vector{Vector{Int}}

Returns minimum cycle basis represented as an array of edge sequence that make up a cycle.

Reference

de Pina, J.C. Applications of shortest path methods. PhD thesis, University of Amsterdam, Nether-lands (1995)

Kavitha, T., Mehlhorn, K., Michail, D. & Paluch, K. E. An $\tilde{O}(m^{2}n)$ Algorithm for Minimum Cycle Basis of Graphs. Algorithmica 52, 333–349 (2008).

Triangles

MolecularGraph.Graph.triangles — Method

triangles(graph::UndirectedGraph) -> Set{Tuple{Int,Int,Int}}

Fast computation of finding triangle node sets in the graph.

Bipartite Graph

MolecularGraph.Graph.twocoloring — Method

twocoloring(graph::UndirectedGraph)

Do 2-coloring and return two sets of nodes.

Maximum cardinality matching

MolecularGraph.Graph.isperfectmatching — Method

isperfectmatching(G::UndirectedGraph) -> Bool
isperfectmatching(G::UndirectedGraph, matching::Set{Int}) -> Bool

Return if the given graph has a perfect matching.

MolecularGraph.Graph.maximummatching — Method

maximummatching(G::UndirectedGraph; method=:Blossom) -> Set{Int}

Compute maximum cardinality matching by Edmonds' blossom algorithm and return the set of matched edges.

Planarity

MolecularGraph.Graph.isouterplanar — Method

isouterplanar(graph::UndirectedGraph) -> Bool

Return whether the graph is outerplanar. The outerplanarity test is based on a planarity test (see isplanar).

MolecularGraph.Graph.isplanar — Method

isplanar(graph::UndirectedGraph) -> Bool

Return whether the graph is planar.

Reference

de Fraysseix, H., & Ossona de Mendez, P. (2012). Trémaux trees and planarity. European Journal of Combinatorics, 33(3), 279–293. https://doi.org/10.1016/j.ejc.2011.09.012

Trémaux trees and planarity. https://arxiv.org/abs/math/0610935

Maximum clique

MolecularGraph.Graph.maximalcliques — Method

maximalcliques(graph::UndirectedGraph; kwargs...) -> FindCliqueState

Return maximal cliques.

Reference

Tomita, E., Tanaka, A., & Takahashi, H. (2006). The worst-case time complexity for generating all maximal cliques and computational experiments. Theoretical Computer Science, 363(1), 28–42. https://doi.org/10.1016/J.TCS.2006.06.015

Cazals, F., & Karande, C. (2008). A note on the problem of reporting maximal cliques. Theoretical Computer Science, 407(1–3), 564–568. https://doi.org/10.1016/j.tcs.2008.05.010

MolecularGraph.Graph.maximalconncliques — Method

maximalconncliques(graph::ModularProduct; kwargs...) -> FindConnCliqueState

Return maximal connected cliques.

Reference

Cazals, F., & Karande, C. (2005). An algorithm for reporting maximal c-cliques. Theoretical Computer Science, 349(3), 484–490. https://doi.org/10.1016/j.tcs.2005.09.038

MolecularGraph.Graph.maximumclique — Method

maximumclique(state::FindCliqueState) -> Set{Int}

Return a maximum clique.

MolecularGraph.Graph.maximumclique — Method

maximumclique(state::FindConnCliqueState) -> Set{Int}

Return a maximum connected clique.

Subgraph match (VF2 algorithm)

MolecularGraph.Graph.edgesubgraph_is_isomorphic — Method

edgesubgraph_is_isomorphic(
    G::UndirectedGraph, H::UndirectedGraph; kwargs...) -> Bool

Return whether an edge-induced subgraph of G is isomorphic to H.

MolecularGraph.Graph.edgesubgraph_isomorphisms — Method

edgesubgraph_isomorphisms(
    G::UndirectedGraph, H::UndirectedGraph;
    nodematcher=(g,h)->true, edgematcher=(g,h)->true,
    kwargs...) -> Iterator

Return an iterator that generate isomorphic mappings between H and edge-induced subgraphs of G. The returned iterator has ig => ih pairs that correspond to the indices of matching edges in G and H, respectively.

nodematcher and edgematcher control the features needed to be counted as a match.

See Graph.edgesubgraph to construct the subgraphs that result from the match.

MolecularGraph.Graph.is_isomorphic — Method

is_isomorphic(G::UndirectedGraph, H::UndirectedGraph; kwargs...) -> Bool

Return whether G and H are isomorphic.

MolecularGraph.Graph.isomorphisms — Method

isomorphisms(G::UndirectedGraph, H::UndirectedGraph; kwargs...) -> Iterator

Return an iterator that generate isomorphic mappings between G and H. The returned iterator has ig => ih pairs that correspond to the indices of matching nodes in G and H, respectively.

MolecularGraph.Graph.nodesubgraph_is_isomorphic — Method

nodesubgraph_is_isomorphic(G::UndirectedGraph, H::UndirectedGraph; kwargs...) -> Bool

Return whether a node-induced subgraph of G is isomorphic to H.

MolecularGraph.Graph.nodesubgraph_isomorphisms — Method

nodesubgraph_isomorphisms(G::UndirectedGraph, H::UndirectedGraph; kwargs...) -> Iterator

Return an iterator that generate isomorphic mappings between H and node-induced subgraphs of G. The returned iterator has ig => ih pairs that correspond to the indices of matching nodes in G and H, respectively.

MolecularGraph.Graph.subgraph_is_monomorphic — Method

subgraph_is_monomorphic(
    G::UndirectedGraph, H::UndirectedGraph; kwargs...) -> Bool

Return whether a subgraph of G is monomorphic to H.

MolecularGraph.Graph.subgraph_monomorphisms — Method

subgraph_monomorphisms(
    G::UndirectedGraph, H::UndirectedGraph; kwargs...) -> Iterator

Generate monomorphism mappings between H and subgraphs of G. The returned iterator has ig => ih pairs that correspond to the indices of matching nodes in G and H, respectively.

MolecularGraph.Graph.VF2Matcher — Type

VF2Matcher{T1<:UndirectedGraph,T2<:UndirectedGraph}(
    G::T1, H::T2, matchtype::Symbol; kwargs...
)

Lazy iterator that generate all isomorphism mappings between G and H.

matchtype should be one of the followings

:isomorphic: G is isomorphic to H

:subgraph_isomorphic: a node-induced subgraph of G is isomorphic to H

:monomorphic: a subgraph of G is monomorphic to H

Options

nodematcher::Function: a function for semantic node attribute matching (default: (a, b) -> true)

edgematcher::Function: a function for semantic edge attribute matching (default: (a, b) -> true)

mandatory::Dict{Int,Int}: mandatory node mapping (if matchtype=:edgeinduced, edge mapping)

forbidden::Dict{Int,Int}: forbidden node mapping (if matchtype=:edgeinduced, edge mapping)

timeout::Union{Int,Nothing}: if specified, abort vf2 calculation when the time reached and return empty iterator (default: 10 seconds)

Maximum common subgraph (Clique detection based algorithm)

MolecularGraph.Graph.maxcommonsubgraph — Method

maxcommonsubgraph(G::UndirectedGraph, H::UndirectedGraph; kwargs...) -> MaxCommonSubgraphResult

Compute maximum common edge induced subgraph between G and H.