Clique

Clique

Clique

MolecularGraph.Graph.maxclique — Method.

maxclique(graph::UDGraph; kwargs...) -> Set{Int}

Compute maximum clique of the graph. For details, see maximalcliques.

source

MolecularGraph.Graph.maximalcliques — Method.

maximalcliques(graph::UDGraph; kwargs...)

Return Channel which generates maximal cliques of the graph. Each cliques are represented as a Set of member nodes.

Reference

Tomita, E., Tanaka, A., & Takahashi, H. (2006). The worst-case time complexity for generating all maximal cliques and computational experiments. Theoretical Computer Science, 363(1), 28–42. https://doi.org/10.1016/J.TCS.2006.06.015

Cazals, F., & Karande, C. (2005). An algorithm for reporting maximal c-cliques. Theoretical Computer Science, 349(3), 484–490. https://doi.org/10.1016/j.tcs.2005.09.038

source