Models

Graph models

Undirected graph

MolecularGraph.MolecularGraphModel.mapgraph — Method.

mapgraph(nodes, edges) -> MapGraph{Node,Edge}

Generate map graph that has given nodes and edges represented by the list of node indices in integer and the list of pairs of node indices, respectively.

source

MolecularGraph.MolecularGraphModel.mapgraph — Method.

mapgraph(graph::UndirectedGraph; clone=false) -> MapGraph

Convert the given graph into a new MapGraph. The node type and edge type are inherited from the original graph. If the given graph is MapGraph, return a copy of the graph.

If you really need mutable nodes and edges, and want the graph with deepcopied elements, implement clone and setnodes as deepcopy methods for nodes and edges, respectively.

source

MolecularGraph.MolecularGraphModel.mapgraph — Method.

mapgraph(::Type{N}, ::Type{E}
    ) where {N<:AbstractNode,E<:UndirectedEdge} -> MapGraph{N,E}()

Generate empty map graph that has nodes and edges with the given types.

source

MolecularGraph.MolecularGraphModel.vectorgraph — Method.

vectorgraph(nodes, edges) -> VectorGraph{Node,Edge}

Generate vector graph that have given nodes and edges represented by the list of node indices in integer and the list of pairs of node indices, respectively.

source

MolecularGraph.MolecularGraphModel.vectorgraph — Method.

vectorgraph(graph::UndirectedGraph; clone=false) -> VectorGraph

Convert the given graph into a new VectorGraph. The node type and edge type are inherited from the original graph.

Node indices are sorted in ascending order and are re-indexed. This behavior is intended for some cannonicalization operations (ex. chirality flag).

If you really need mutable nodes and edges, and want the graph with deepcopied elements, implement clone and setnodes as deepcopy methods for nodes and edges, respectively.

source

MolecularGraph.MolecularGraphModel.vectorgraph — Method.

vectorgraph(::Type{N}, ::Type{E}
    ) where {N<:AbstractNode,E<:UndirectedEdge} -> VectorGraph{N,E}()

Generate empty vector graph that have nodes and edges with the given types.

source

Directed graph

MolecularGraph.MolecularGraphModel.mapdigraph — Method.

mapdigraph(nodes, edges) -> MapDiGraph{Node,Arrow}

Generate MapDiGraph that have given nodes and edges represented by the list of node indices in integer and the list of pairs of node indices, respectively.

source

MolecularGraph.MolecularGraphModel.mapdigraph — Method.

mapdigraph(graph::DirectedGraph; clone=false) -> MapDiGraph

Convert the given graph into a new MapDiGraph. The node type and edge type are inherited from the original graph. If the given graph is MapDiGraph, return a copy of the graph.

source

MolecularGraph.MolecularGraphModel.mapdigraph — Method.

mapdigraph(::Type{N}, ::Type{E}
    ) where {N<:AbstractNode,E<:DirectedEdge} -> MapDiGraph{N,E}()

Generate empty MapDiGraph that have nodes and edges with the given types.

source

Multi graph

MolecularGraph.MolecularGraphModel.mapmultigraph — Method.

mapmultigraph(nodes, edges) -> MapMultiGraph{Node,Edge}

Generate MapMultiGraph that have given nodes and edges represented by the list of node indices in integer and the list of pairs of node indices, respectively.

source

MolecularGraph.MolecularGraphModel.mapmultigraph — Method.

mapmultigraph(::Type{N}, ::Type{E}
    ) where {N<:AbstractNode,E<:UndirectedEdge} -> MapMultiGraph{N,E}()

Generate empty MapMultiGraph that have nodes and edges with the given types.

source