Operations

Graph operations

Disjoint union

MolecularGraph.MolecularGraphModel.disjointunion! — Method.

disjointunion!(g1::T, g2::T, G::T...) where {T<:OrderedGraph} -> T

Generate disjoint union graph of given graphs. g1 will be overwritten by the union graph. Unlike non-destructive disjointunion, g1 does not retain any information about other given graphs but a bit faster.

source

MolecularGraph.MolecularGraphModel.disjointunion — Method.

disjointunion(g1::UndirectedGraph, g2::UndirectedGraph,
    G::UndirectedGraph...) -> DisjointUnionGraph

Generate disjoint union graph of given graphs. The new graph with type DisjointUnionGraph retains mapping to the original graphs as nodes and edges attributes.

source

Line graph

MolecularGraph.MolecularGraphModel.linegraph — Method.

linegraph(G::AbstractGraph) -> LineGraph

Generate line graph.

source

Product of graphs

MolecularGraph.MolecularGraphModel.cartesianproduct — Method.

cartesianproduct(G::OrderedGraph, H::OrderedGraph) -> CartesianProduct

Return the cartesian product of graphs G and H.

source

MolecularGraph.MolecularGraphModel.modularproduct — Method.

modularproduct(G::OrderedGraph, H::OrderedGraph) -> ModularProduct

Return the modular product of graphs G and H.

source