Shortest path

Shortest path

Shortest path

MolecularGraph.MolecularGraphModel.diameter — Method.

diameter(graph::UndirectedGraph) => Int

Compute the diameter of the graph (the largest eccentricity of any nodes).

source

MolecularGraph.MolecularGraphModel.distance — Method.

distance(graph::UndirectedGraph, root) -> Dict{Int,Union{Int,Nothing}}

Compute the distance from root to any other nodes. If the nodes are not reachable each other, the value will be nothing.

source

MolecularGraph.MolecularGraphModel.distancematrix — Method.

distancematrix(graph::UndirectedGraph) -> Matrix{Float64}

Compute the distance among each other nodes.

Note that the type of the generated matrix will be Float64. If the nodes are not reachable each other, the distance value will be Inf.

source

MolecularGraph.MolecularGraphModel.eccentricity — Method.

eccentricity(graph::UndirectedGraph, v) => Int

Compute the eccentricity of the graph (the largest distance between v and any other nodes).

source

MolecularGraph.MolecularGraphModel.longestshortestpath — Method.

longestshortestpath(graph::UndirectedGraph) -> Vector{Int}

Compute the longest shortest path in the graph (a path between two arbitrary peripheral nodes) as a vector of nodes that starts with one of the peripheral node and ends with the other side.

source

MolecularGraph.MolecularGraphModel.shortestpath — Method.

shortestpath(graph::UndirectedGraph, u, v) -> Vector{Int}

Compute the shortest path between u and v as a vector of the nodes that starts with u and ends withv. Return nothing if u == v or not reachable.

source