Graph topology

Cycle

Connectivity

Planarity

MolecularGraph.Graph.isouterplanar — Method

isouterplanar(graph::UndirectedGraph) -> Bool

Return whether the graph is outerplanar. The outerplanarity test is based on a planarity test (see isplanar).

MolecularGraph.Graph.isplanar — Method

isplanar(graph::UndirectedGraph) -> Bool

Return whether the graph is planar.

Reference

de Fraysseix, H., & Ossona de Mendez, P. (2012). Trémaux trees and planarity. European Journal of Combinatorics, 33(3), 279–293. https://doi.org/10.1016/j.ejc.2011.09.012

Triangles

MolecularGraph.Graph.triangles — Method

triangles(graph::UndirectedGraph) -> Set{Tuple{Int,Int,Int}}

Fast computation of finding triangle node sets in the graph.