Traversal

Graph traversal

Shortest path

MolecularGraph.MolecularGraphModel.diameter — Method.

diameter(graph::AbstractGraph) -> Int

Compute the diameter of the graph (the largest eccentricity of any nodes).

source

MolecularGraph.MolecularGraphModel.distance — Method.

distance(graph::AbstractGraph, source::Int, target::Int) -> Int
reversedistance(graph::DirectedGraph, source::Int, target::Int) -> Int

Compute the distance (shortest path length) from source to target. If the nodes are not reachable each other, the value will be nothing.

source

MolecularGraph.MolecularGraphModel.distancematrix — Method.

distancematrix(graph::OrderedGraph) -> Matrix{Float64}

Generate the distance matrix of the graph.

Note that the type of the generated matrix will be Float64. If the nodes are not reachable each other, the distance value will be Inf.

source

MolecularGraph.MolecularGraphModel.eccentricity — Method.

eccentricity(graph::UndirectedGraph, v::Int) -> Int

Compute the eccentricity of the graph (the largest distance between v and any other nodes).

source

MolecularGraph.MolecularGraphModel.isreachable — Method.

reachablenodes(graph::AbstractGraph, u::Int, v::Int) -> Bool
reversereachablenodes(graph::DirectedGraph, u::Int, v::Int) -> Bool

Return whether the node v is reachable from u.

source

MolecularGraph.MolecularGraphModel.longestshortestpathnodes — Method.

longestshortestpathnodes(graph::UndirectedGraph) -> Vector{Int}

Compute the longest shortest path in the graph (a path between two arbitrary peripheral nodes) as a vector of nodes that starts with one of the peripheral node and ends with the other side.

source

MolecularGraph.MolecularGraphModel.reachablenodes — Method.

reachablenodes(graph::AbstractGraph, node::Int) -> Set{Int}
reversereachablenodes(graph::DirectedGraph, node::Int) -> Set{Int}

Return the set of reachable nodes from node.

source

MolecularGraph.MolecularGraphModel.shortestpathedges — Method.

shortestpathedges(graph::UndirectedGraph, u::Int, v::Int) -> Vector{Int}

Compute the shortest path between u and v as a vector of the edges that forms the path. Return nothing if not reachable.

source

MolecularGraph.MolecularGraphModel.shortestpathnodes — Method.

shortestpathnodes(graph::UndirectedGraph, u::Int, v::Int) -> Vector{Int}

Compute the shortest path between u and v as a vector of the nodes that forms the path. Return nothing if not reachable.

source