InChI

MolecularGraph.inchi — Method

inchi(molblock::String; options::String = "", verbose::Bool = false) -> Union{String,Nothing}
inchi(mol::MolGraph; options::String = "", verbose::Bool = false) -> Union{String,Nothing}

Generate InChI string from molblock string or molecule.

Options, e.g. "SNon" for 'no stereo information' are specified in https://github.com/mojaie/libinchi/blob/master/INCHIBASE/src/inchiapi.h

MolecularGraph.inchikey — Method

inchikey(inchi::String) -> Union{String,Nothing}
inchikey(mol::MolGraph) -> Union{String,Nothing}

Generate InChI key from InChI string or molecule

MolecularGraph.inchitomol — Method

function inchitomol(inchi::String; options = "", verbose = false)

Generate molecule from inchi string, options are specified in https://github.com/mojaie/libinchi/blob/master/INCHIBASE/src/inchiapi.h

MolecularGraph.inchitosdf — Method

inchitosdf(inchi::String; options::String = "") -> Union{String,Nothing}

Generate sdf string from inchi string, options are specified in https://github.com/mojaie/libinchi/blob/master/INCHIBASE/src/inchiapi.h