Molecular structure drawing

MolecularGraph.atomnumber — Method

atomnumber(atomsymbol::Symbol) -> Int
atomnumber(atom::Atom) -> Int

Return atom number.

MolecularGraph.atomsymbol — Method

atomsymbol(n::Int) -> Symbol

Return atom symbol of given atomic number.

MolecularGraph.graphmol — Method

graphmol(mol::SubgraphView{GraphMol}) -> GraphMol

Generate a new GraphMol from a substructure view.

Graph property caches and attributes are not inherited.

MolecularGraph.graphmol — Method

graphmol(atoms::Vector{Atom}, bonds::Vector{Bond}) -> GraphMol

Generate GraphMol that has the given atom objects and edge objects.

MolecularGraph.graphmol — Method

graphmol() -> GraphMol

Generate empty GraphMol.

MolecularGraph.graphmol — Method

graphmol(mol::GraphMol; clearcache=true) -> GraphMol

Copy GraphMol. Cached properties will be removed.

MolecularGraph.querymol — Method

querymol(atoms::Vector{Atom}, bonds::Vector{Bond}) -> GraphMol

Generate QueryMol that has the given atom objects and edge objects.

MolecularGraph.querymol — Method

querymol() -> QueryMol

Generate empty QueryMol.

MolecularGraph.todict — Method

todict(graph::GraphMol) -> Dict{String,Any}

Convert molecule object into JSON compatible dictionary.