Stereochemistry

MolecularGraph.remove_stereo_hydrogen! — Method

remove_stereo_hydrogen!(mol::SimpleMolGraph, v::Integer) -> Bool

Safely remove explicit hydrogens connected to stereocenter node v and return the vertex index of the removed hydrogen.

MolecularGraph.set_stereobond! — Method

set_stereocenter!(mol::SimpleMolGraph, bond, v1, v2, is_cis) -> Nothing

Set stereocenter information to graph properties.

MolecularGraph.set_stereocenter! — Method

set_stereocenter!(mol::SimpleMolGraph, center, looking_from, v1, v2, is_clockwise) -> Nothing

Set stereocenter information to graph properties.

MolecularGraph.stereobond_from_sdf2d! — Method

stereobond_from_sdf2d!(mol::MolGraph) -> Nothing

Set cis-trans diastereomerism information obtained from 2D SDFile.

MolecularGraph.stereobond_from_smiles! — Method

stereobond_from_smiles!(mol::MolGraph) -> Nothing

Set cis-trans diastereomerism information obtained from SMILES.

MolecularGraph.stereocenter_from_sdf2d! — Method

stereocenter_from_sdf2d!(mol::MolGraph) -> Nothing

Set stereocenter information obtained from 2D SDFile.

MolecularGraph.stereocenter_from_smiles! — Method

stereocenter_from_smiles!(mol::MolGraph) -> Nothing

Set stereocenter information obtained from SMILES.