Structure match

MolecularGraph.atommatch — Method

atommatch(mol1::UndirectedGraph, mol2::UndirectedGraph) -> Function

Return a default atom attribute comparator between two atoms.

MolecularGraph.atommatch — Method

atommatch(mol::UndirectedGraph, querymol::QueryMol) -> Function

Return a default atom attribute comparator that returns true if the atom satisfies all the queryatom conditions.

MolecularGraph.atommatch — Method

atommatch(qmol1::QueryMol, qmol2::QueryMol) -> Function

Return an atom attribute comparator that returns true if a2 contains a1.

MolecularGraph.bondmatch — Method

bondmatch(mol1::UndirectedGraph, mol2::UndirectedGraph) -> Function

Return a default bond attribute comparator between two bonds.

MolecularGraph.bondmatch — Method

bondmatch(mol::UndirectedGraph, querymol::QueryMol) -> Function

Return a default bond attribute comparator that returns true if the bond satisfies all the querybond conditions.

MolecularGraph.bondmatch — Method

bondmatch(qmol1::QueryMol, qmol2::QueryMol) -> Function

Return a bond attribute comparator that returns true if b2 contains b1.

MolecularGraph.edgeinducedmatches — Method

edgeinducedmatches(
    mol::UndirectedGraph, query::UndirectedGraph; kwargs...) -> Iterator

Return a lazy iterator that generate edge mappings between mol and query if a edge-induced subgraph of mol graph is isomorphic to query graph. See MolecularGraph.structmatches for available options.

MolecularGraph.exactatommatch — Method

exactatommatch(qmol1::QueryMol, qmol2::QueryMol) -> Function

Return a default atom attribute comparator between two atom queries.

MolecularGraph.exactbondmatch — Method

exactbondmatch(qmol1::QueryMol, qmol2::QueryMol) -> Function

Return a default bond attribute comparator between two bond queries.

MolecularGraph.exactmatches — Method

exactmatches(
    mol::UndirectedGraph, query::UndirectedGraph; kwargs...) -> Iterator

Return a lazy iterator that generate node mappings between mol and query if these are exactly same. See MolecularGraph.structmatches for available options.

MolecularGraph.hasedgeinducedmatch — Method

hasedgeinducedmatch(
    mol::UndirectedGraph, query::UndirectedGraph; kwargs...) -> Iterator

Return whether a edge-induced subgraph of mol graph is isomorphic to query graph. See MolecularGraph.structmatches for available options.

MolecularGraph.hasexactmatch — Method

hasexactmatch(
    mol::UndirectedGraph, query::UndirectedGraph; kwargs...) -> Iterator

Return whether mol and query have exactly the same structure. See MolecularGraph.structmatches for available options.

MolecularGraph.hasnodeinducedmatch — Method

hasnodeinducedmatch(
    mol::UndirectedGraph, query::UndirectedGraph; kwargs...) -> Iterator

Return whether a node-induced subgraph of mol graph is isomorphic to query graph. See MolecularGraph.structmatches for available options.

MolecularGraph.hassubstructmatch — Method

hassubstructmatch(
    mol::UndirectedGraph, query::UndirectedGraph; kwargs...) -> Iterator

Return whether mol has query as a substructure. See MolecularGraph.structmatches for available options.

MolecularGraph.maxcommonsubstruct — Method

maxcommonsubstruct(mol1::UndirectedGraph, mol2::UndirectedGraph, matchtype; kwargs...) -> MaxCommonSubgraphResult

Compute maximum common substructure (MCS) of mol1 and mol2.

Keyword arguments

matchtype(Symbol): :nodeinduced or :edgeinduced (default=:edgeinduced).

connected(Bool): if true, apply connected MCS constraint (default=false).

topological(Bool): if true, apply topological constraint (default=false).

diameter(Int): distance cutoff for topological constraint.

tolerance(Int): distance mismatch tolerance for topological constraint.

timeout(Int): abort calculation and return suboptimal results if the execution

time has reached the given value (default=60, in seconds).

targetsize(Int): abort calculation and return suboptimal result so far if the

given mcs size achieved.

References

Kawabata, T. (2011). Build-Up Algorithm for Atomic Correspondence between

Chemical Structures. Journal of Chemical Information and Modeling, 51(8), 1775–1787. https://doi.org/10.1021/ci2001023

https://www.jstage.jst.go.jp/article/ciqs/2017/0/2017P4/article/-char/en

MolecularGraph.nodeinducedmatches — Method

nodeinducedmatches(
    mol::UndirectedGraph, query::UndirectedGraph; kwargs...) -> Iterator

Return a lazy iterator that generate node mappings between mol and query if a node-induced subgraph of mol graph is isomorphic to query graph. See MolecularGraph.structmatches for available options.

MolecularGraph.structmatches — Method

structmatches(
    mol::UndirectedGraph, query::UndirectedGraph,
    matchtype::Symbol; kwargs...) -> Iterator

Return a lazy iterator that generate all isomorphism mappings between mol and query.

matchtype should be one of the followings

:exact: mol is isomorphic to query

:substruct: a subgraph of mol is monomorphic to query (typical substructure search)

:nodeinduced: a node-induced subgraph of mol is isomorphic to query

:edgeinduced: an edge-induced subgraph of mol is isomorphic to query`

options

prefilter::Bool: if true, apply simple prefilter by graph size and topology to skip vf2 calculation (dafault: true)

fastsingleton::Bool: if true, skip vf2 if the query is single node or edge that can be matched by simple iteration (dafault: false)

atommatcher::Function: a function for semantic atom attribute matching (default: MolecularGraph.atommatch)

bondmatcher::Function: a function for semantic bond attribute matching (default: MolecularGraph.edgematch)

mandatory::Dict{Int,Int}: mandatory node mapping (or edge mapping if matchtype=:edgeinduced)

forbidden::Dict{Int,Int}: forbidden node mapping (or edge mapping if matchtype=:edgeinduced)

timeout::Union{Int,Nothing}: if specified, abort vf2 calculation when the time reached and return empty iterator (default: 10 seconds).

Note that null mol and null query never match (e.g. isstructmatch(smilestomol(""), smilestomol("")) is false)

MolecularGraph.substructmatches — Method

substructmatches(
    mol::UndirectedGraph, query::UndirectedGraph; kwargs...) -> Iterator

Return a lazy iterator that generate node mappings between mol and query if mol has query as a substructure. See MolecularGraph.structmatches for available options.