Maximum clique

MolecularGraph.Graph.maximalcliques — Method

maximalcliques(graph::UndirectedGraph; kwargs...
    ) -> Tuple{Vector{Set{Int}}, Symbol}

Return maximal cliques.

Reference

Tomita, E., Tanaka, A., & Takahashi, H. (2006). The worst-case time complexity for generating all maximal cliques and computational experiments. Theoretical Computer Science, 363(1), 28–42. https://doi.org/10.1016/J.TCS.2006.06.015

Cazals, F., & Karande, C. (2008). A note on the problem of reporting maximal cliques. Theoretical Computer Science, 407(1–3), 564–568. https://doi.org/10.1016/j.tcs.2008.05.010

MolecularGraph.Graph.maximalconncliques — Method

maximalconncliques(graph::ModularProduct; kwargs...
    ) -> Tuple{Vector{Set{Int}}, Symbol}

Return maximal connected cliques.

Reference

Cazals, F., & Karande, C. (2005). An algorithm for reporting maximal c-cliques. Theoretical Computer Science, 349(3), 484–490. https://doi.org/10.1016/j.tcs.2005.09.038

MolecularGraph.Graph.maximumclique — Method

maximumclique(graph::UndirectedGraph; kwargs...) -> Tuple{Set{Int}, Symbol}

Return a maximum clique.

MolecularGraph.Graph.maximumconnclique — Method

maximumconnclique(graph::ModularProduct; kwargs...
    ) -> Tuple{Set{Int}, Symbol}

Return a maximum connected clique.