Interface

Interface

Interface

MolecularGraph.MolecularGraphModel.edgecount — Function.

edgecount(graph) -> Int

Return the number of graph edges.

source

MolecularGraph.MolecularGraphModel.edgekeys — Function.

edgekeys(graph) -> Vector{Int}

Return graph edge keys. If the given graph is a vector graph, the keys are in ascending order, whereas the order of indices in map graph is not guaranteed.

source

MolecularGraph.MolecularGraphModel.edgeset — Function.

edgeset(graph) -> Set{Int}

Return the set of edge keys.

source

MolecularGraph.MolecularGraphModel.edgesiter — Function.

edgesiter(graph)

An iterator that yields (i, e) where i is the edge index, and e is the edge object at the index i within the graph.

source

MolecularGraph.MolecularGraphModel.edgetype — Function.

edgetype(graph)

Return the edge type of the graph

source

MolecularGraph.MolecularGraphModel.getedge — Method.

getedge(graph, u, v) -> AbstractEdge

Retrieve an edge object which connects the given nodes.

source

MolecularGraph.MolecularGraphModel.getedge — Method.

getedge(graph, index) -> AbstractEdge

Retrieve the edge object at the given index within the graph.

source

MolecularGraph.MolecularGraphModel.getnode — Function.

getnode(graph, index) -> AbstractNode

Retrieve the node object at the given index within the graph.

source

MolecularGraph.MolecularGraphModel.hasedge — Function.

hasedge(graph, u, v) -> AbstractEdge

Return whether the given two nodes are connected by at least one edge.

source

MolecularGraph.MolecularGraphModel.indegree — Function.

indegree(graph, n) -> Int

Return the number of inneighbors of the node 'n'.

source

MolecularGraph.MolecularGraphModel.inneighbors — Function.

inneighbors(graph, n) -> Dict{Int,Int}

Return the mapping of predecessor node keys and in edge keys connected to the given node.

source

MolecularGraph.MolecularGraphModel.neighborcount — Function.

neighborcount(graph, n) -> Int
degree(graph, n) -> Int

Return the number of adjacent nodes of the node 'n'.

source

MolecularGraph.MolecularGraphModel.neighbors — Function.

neighbors(graph, n) -> Dict{Int,Int}

Return the mapping of adjacent node keys and incident edge keys connected to the given node. If the graph is directed graph, both outneighbors and inneighbors are mapped.

source

MolecularGraph.MolecularGraphModel.nodecount — Function.

nodecount(graph) -> Int

Return the number of graph nodes.

source

MolecularGraph.MolecularGraphModel.nodekeys — Function.

nodekeys(graph) -> Vector{Int}

Return graph node keys. If the given graph is a vector graph, the keys are in ascending order, whereas the order of indices in map graph is not guaranteed.

source

MolecularGraph.MolecularGraphModel.nodeset — Function.

nodeset(graph) -> Set{Int}

Return the set of node keys.

source

MolecularGraph.MolecularGraphModel.nodesiter — Function.

nodesiter(graph)

An iterator that yields (i, n) where i is the node index, and n is the node object at the index i within the graph.

source

MolecularGraph.MolecularGraphModel.nodetype — Function.

nodetype(graph)

Return the node type of the graph

source

MolecularGraph.MolecularGraphModel.outdegree — Function.

outdegree(graph, n) -> Int

Return the number of outneighbors of the node 'n'.

source

MolecularGraph.MolecularGraphModel.outneighbors — Function.

outneighbors(graph, n) -> Dict{Int,Int}

Return the mapping of successor node keys and out edge keys connected to the given node.

source

MolecularGraph.MolecularGraphModel.unlinkedge! — Method.

unlinkedge!(graph, u, v)

Delete the edge that connect nodes u and v.

source

MolecularGraph.MolecularGraphModel.unlinkedge! — Method.

unlinkedge!(graph, e)

Delete the edge at the index of e.

source

MolecularGraph.MolecularGraphModel.unlinknode! — Function.

unlinknode!(graph, n)

Delete the node at the index of n and its incident edges.

source

MolecularGraph.MolecularGraphModel.updateedge! — Method.

updateedge!(graph, edge, u, v)

Rebind the edge that connects nodes u and v by the given edge object. If the nodes do not exist, throws KeyError.

source

MolecularGraph.MolecularGraphModel.updateedge! — Method.

updateedge!(graph, edge, e)

Rebind the edge object stored at e of the graph by the given edge object. If the index does not exist, add new edge to the position e.

source

MolecularGraph.MolecularGraphModel.updatenode! — Function.

updatenode!(graph, node, n)

Rebind the node object stored at n of the graph by the given node object. If the index does not exist, add new node to the position n.

source