Interface

Interface

Interface

MolecularGraph.MolecularGraphModel.addnode! — Method.

addnode!(graph) -> Int
addnode!(graph, attr) -> Int

Add new node and return the node index. If the node attribute type is required, specify the node attribute object by node keyword.

source

MolecularGraph.MolecularGraphModel.edgeattr — Method.

edgeattr(graph::AbstractGraph, u::Int, v::Int
    ) -> Union{AbstractEdge,Nothing}

Return the attribute object of an edge that connects u and v. If not found, return nothing.

source

MolecularGraph.MolecularGraphModel.edgeattr — Method.

edgeattr(graph::AbstractGraph, i::Int) -> AbstractEdge

Return the attribute object of edge i.

source

MolecularGraph.MolecularGraphModel.edgeattrs — Method.

edgeattrs(graph::Union{OrderedGraph,OrderedDiGraph}) -> Vector{AbstractEdge}

Return graph edge attributes.

source

MolecularGraph.MolecularGraphModel.edgeattrtype — Method.

edgeattrtype(graph) -> Type

Return edge attribute type.

source

MolecularGraph.MolecularGraphModel.edgecount — Method.

edgecount(graph::AbstractGraph) -> Int

Return the number of graph edges.

source

MolecularGraph.MolecularGraphModel.edgeset — Method.

edgeset(graph::Union{OrderedGraph,OrderedDiGraph}) -> Set{Int}

Return the set of edge keys.

source

MolecularGraph.MolecularGraphModel.getedge — Method.

getedge(graph::AbstractGraph, i::Int) -> Tuple{Int,Int}
getedge(graph::HyperGraph, i::Int) -> Set{Int}

Return an edge.

source

MolecularGraph.MolecularGraphModel.hasedge — Method.

hasedge(graph::UndirectedGraph, u::Int, v::Int) -> Bool
hasedge(graph::DirectedGraph, source::Int, target::Int) -> Bool

Return whether the given two nodes are connected by at least one edge.

source

MolecularGraph.MolecularGraphModel.indegree — Method.

indegree(graph::DirectedGraph, n::Int) -> Int

Return the number of inneighbors of the node 'n'.

source

MolecularGraph.MolecularGraphModel.inneighbors — Method.

inneighbors(graph::DirectedGraph, i::Int) -> Dict{Int,Int}

Return the mapping of predecessor node keys and in edge keys connected to the given node.

source

MolecularGraph.MolecularGraphModel.neighborcount — Method.

neighborcount(graph::AbstractGraph, n::Int) -> Int
degree(graph::AbstractGraph, n::Int) -> Int

Return the number of adjacent nodes of the node 'n'.

source

MolecularGraph.MolecularGraphModel.neighbors — Method.

neighbors(graph, i) -> Dict{Int,Int}

Return the mapping of incident edges and adjacent nodes of node i. If the graph is directed graph, both outneighbors and inneighbors are mapped.

source

MolecularGraph.MolecularGraphModel.nodeattr — Method.

nodeattr(graph::AbstractGraph, i::Int) -> AbstractNode

Return the attribute object of node i.

source

MolecularGraph.MolecularGraphModel.nodeattrs — Method.

nodeattrs(graph::Union{OrderedGraph,OrderedDiGraph}) -> Vector{AbstractNode}

Return graph node attributes.

source

MolecularGraph.MolecularGraphModel.nodeattrtype — Method.

nodeattrtype(graph) -> Type

Return node attribute type.

source

MolecularGraph.MolecularGraphModel.nodecount — Method.

nodecount(graph::AbstractGraph) -> Int

Return the number of graph nodes.

source

MolecularGraph.MolecularGraphModel.nodeset — Method.

nodeset(graph::Union{OrderedGraph,OrderedDiGraph}) -> Set{Int}

Return the set of node keys.

source

MolecularGraph.MolecularGraphModel.outdegree — Method.

outdegree(graph::DirectedGraph, n::Int) -> Int

Return the number of outneighbors of the node 'n'.

source

MolecularGraph.MolecularGraphModel.outneighbors — Method.

outneighbors(graph::DirectedGraph, i::Int) -> Dict{Int,Int}

Return the mapping of successor node keys and out edge keys connected to the given node.

source

MolecularGraph.MolecularGraphModel.setedgeattr! — Method.

setedgeattr!(graph::AbstractGraph, i::Int, attr::AbstractNode)

Update the edge attribute.

source

MolecularGraph.MolecularGraphModel.setnodeattr! — Method.

setnodeattr!(graph::AbstractGraph, i::Int, attr::AbstractNode)

Update the node attribute.

source

MolecularGraph.MolecularGraphModel.unlinkedges — Function.

unlinkedges(graph, edges)

Delete given edges from the graph.

source

MolecularGraph.MolecularGraphModel.unlinknodes — Function.

unlinknodes(graph, nodes)

Delete given nodes and its incident edges from the graph.

source