Isomorphism

Graph isomorphism

Subgraph match (VF2 algorithm)

MolecularGraph.MolecularGraphModel.edgesubgraphmatch — Method.

edgesubgraphmatch(
    G::AbstractGraph, H::AbstractGraph; kwargs...) -> Dict{Int,Int}

Return a edge induced subgraph isomorphism mapping between G and H. If no match found, return nothing.

source

MolecularGraph.MolecularGraphModel.edgesubgraphmatches — Method.

edgesubgraphmatch(G::AbstractGraph, H::AbstractGraph; kwargs...) -> Iterator

Generate edge induced subgraph isomorphism mappings between G and H.

source

MolecularGraph.MolecularGraphModel.graphmatch — Method.

graphmatch(G::AbstractGraph, H::AbstractGraph; kwargs...) -> Dict{Int,Int}

Return an isomorphism mapping between G and H. If no match found, return nothing.

source

MolecularGraph.MolecularGraphModel.graphmatches — Method.

graphmatches(G::AbstractGraph, H::AbstractGraph; kwargs...) -> Iterator

Generate isomorphism mappings between G and H. If no match found, return nothing.

source

MolecularGraph.MolecularGraphModel.isedgesubgraphmatch — Method.

isedgesubgraphmatch(G::AbstractGraph, H::AbstractGraph; kwargs...) -> Bool

Return true if a node induced subgraph of G and H are isomorphic.

source

MolecularGraph.MolecularGraphModel.isgraphmatch — Method.

isgraphmatch(G::AbstractGraph, H::AbstractGraph; kwargs...) -> Bool

Return true if G and H are isomorphic.

source

MolecularGraph.MolecularGraphModel.issubgraphmatch — Method.

issubgraphmatch(G::AbstractGraph, H::AbstractGraph; kwargs...) -> Bool

Return true if a node induced subgraph of G and H are isomorphic.

source

MolecularGraph.MolecularGraphModel.subgraphmatch — Method.

subgraphmatch(
    G::AbstractGraph, H::AbstractGraph; kwargs...) -> Dict{Int,Int}

Return a subgraph isomorphism mapping between G and H. If no match found, return nothing.

source

MolecularGraph.MolecularGraphModel.subgraphmatches — Method.

subgraphmatches(G::AbstractGraph, H::AbstractGraph; kwargs...) -> Iterator

Generate subgraph isomorphism mappings between G and H.

Keyword arguments

nodematcher(Function): node matcher function that takes two node indices as

arguments.

edgematcher(Function): edge matcher function that takes two edge indices as

arguments.

mandatory(Dict{Int,Int}): mandatory node matches (available for only VF2)

forbidden(Dict{Int,Int}): forbidden node matches (available for only VF2)

source

Maximum common subgraph (Clique detection based algorithm)

MolecularGraph.MolecularGraphModel.findmces — Method.

findmces(G::UndirectedGraph, H::UndirectedGraph; kwargs...) -> MCSResult

Compute maximum common edge induced subgraph between G and H.

source

MolecularGraph.MolecularGraphModel.findmcis — Method.

findmcis(G::UndirectedGraph, H::UndirectedGraph; kwargs...) -> MCSResult

Compute maximum common induced subgraph between G and H.

Keyword arguments:

connectivity: :Connected, :Disconnected (for mode=:MCS)

constraint: TopologicalConstraint, DiameterRestriction (for mode=:MCS)

timeout(Int): return suboptimal result so far if the execution time exceeded

the given value (in second).

threshold(Int): return suboptimal result so far if the given mcs size

achieved.

nodematcher(Function): node matcher function that takes two node indices as

arguments.

edgematcher(Function): edge matcher function that takes two edge indices as

arguments.

theta(Int): distance mismatch tolerance in topologically constrainted MCS

diameter(Int): diameter size in MCS with diameter restriction

source

MolecularGraph.MolecularGraphModel.mcessize — Method.

mcessize(G::AbstractGraph, H::AbstractGraph; kwargs...) -> Int

Return the maximum common edge induced subgraph size (number of edges).

source

MolecularGraph.MolecularGraphModel.mcissize — Method.

mcissize(G::AbstractGraph, H::AbstractGraph; kwargs...) -> Int

Return the maximum common induced subgraph size (number of nodes).

source