Graph topology

Cycle

MolecularGraph.Graph.edgemincycles — Method

edgemincycles(graph::UndirectedGraph) -> Vector{Vector{Int}}

Calculate minimum cycle basis of the unweighted undirected graph by de Pina algorithm re-interpreted by Kavitha et al. This returns an array of edge sets that consist of the cycles.

MolecularGraph.Graph.mincycles — Method

mincycles(graph::UndirectedGraph) -> Vector{Vector{Int}}

Return an array of node sets that consist of the minimum cycle basis.

Connectivity

MolecularGraph.Graph.bridges — Method

bridges(graph::UndirectedGraph) -> Set{Int}

Compute biconnectivity and return bridges.

MolecularGraph.Graph.connectedcomponents — Method

connectedcomponents(graph::UndirectedGraph) -> Vector{Set{Int}}

Compute connectivity and return sets of the connected components.

MolecularGraph.Graph.connectedmembership — Method

connectedmembership(graph::OrderedGraph) -> Vector{Int}

Return connected component membership array.

MolecularGraph.Graph.cutvertices — Method

cutvertices(graph::UndirectedGraph) -> Set{Int}

Compute biconnectivity and return cut vertices (articulation points).

MolecularGraph.Graph.edgebiconnectedcomponents — Method

edgebiconnectedcomponents(graph::UndirectedGraph) -> Vector{Vector{Int}}

Compute sets of biconnected component edges.

Planarity

MolecularGraph.Graph.isouterplanar — Method

isouterplanar(graph::UndirectedGraph) -> Bool

Return whether the graph is outerplanar. The outerplanarity test is based on a planarity test (see isplanar).

MolecularGraph.Graph.isplanar — Method

isplanar(graph::UndirectedGraph) -> Bool

Return whether the graph is planar.

Reference

de Fraysseix, H., & Ossona de Mendez, P. (2012). Trémaux trees and planarity. European Journal of Combinatorics, 33(3), 279–293. https://doi.org/10.1016/j.ejc.2011.09.012

Triangles

MolecularGraph.Graph.triangles — Method

triangles(graph::UndirectedGraph) -> Set{Tuple{Int,Int,Int}}

Fast computation of finding triangle node sets in the graph.