Molecular structure drawing

MolecularGraph.atom_color — Method

atomcolor(mol::SimpleMolGraph; setting=DRAW_SETTING) -> Vector{Color}

Return atom colors for molecule 2D drawing

MolecularGraph.chargesign — Method

chargesign(charge::Int) -> String

Get a charge sign.

MolecularGraph.draw2d! — Method

draw2d!(canvas::Canvas, mol::UndirectedGraph; kwargs...)

Draw molecular image to the canvas.

MolecularGraph.is_atom_visible — Method

is_atom_visible(mol::SimpleMolGraph; setting=DRAW_SETTING) -> Vector{Bool}

Return whether the atom is visible in the 2D drawing. show_carbon: simple - does not show carbon labels, terminal - show only terminal carbonlabels, all - show all carbon labels.

MolecularGraph.ballstick — Method

ballstick(mol::UndirectedGraph; radii=0.3, bonddiameter=0.1)

Represent mol as a ball-and-stick model in three dimensions. mol should have 3d atom positions represented in Angstroms. 3D SDF files can be downloaded from sites such as PubChem.

radii optionally specifies the radii of the balls, in Angstroms. bonddiameter optionally specifies the radii of the sticks, in Angstroms.

This function requires that you load one of the backends of the Makie/GLMakie/CairoMakie family.

MolecularGraph.spacefilling — Method

spacefilling(mol::UndirectedGraph; radii="van der Waals")

Represent mol as a space-filling (Calotte) model in three dimensions. mol should have 3d atom positions represented in Angstroms. (3D SDF files can be downloaded from sites such as PubChem.) The two supported options for radii are "van der Waals" and "covalent"; the former are available only for main-group elements, and the latter are available for all.

This function requires that you load one of the backends of the Makie/GLMakie/CairoMakie family.

MolecularGraph.stick — Method

stick(mol::UndirectedGraph; size=0.3)

Represent mol as a stick model in three dimensions. mol should have 3d atom positions represented in Angstroms. 3D SDF files can be downloaded from sites such as PubChem.

size optionally specifies the width of the sticks, in Angstroms.

This function requires that you load one of the backends of the Makie/GLMakie/CairoMakie family.

MolecularGraph.wire — Method

wire(mol::UndirectedGraph; size=0.1)

Represent mol as a wire-frame model in three dimensions. mol should have 3d atom positions represented in Angstroms. 3D SDF files can be downloaded from sites such as PubChem.

size optionally specifies the width of the bonds, in Angstroms.

This function requires that you load one of the backends of the Makie/GLMakie/CairoMakie family.

MolecularGraph.drawsvg — Method

drawsvg(mol::SimpleMolGraph, width::Int, height::Int)

Generate molecular structure image as a SVG format string.

width and height specifies the size of the image (width and height attribute of svg tag).

MolecularGraph.html_fixed_size — Method

html_fixed_size(svg, width, height)

Generate fixed-size HTML wrapper for the SVG element.

MolecularGraph.html_grid — Method

html_grid(svgs, cols, rowheight)

Generate grid layout HTML wrapper for the SVG elements.

The number of columns and grid row height in pixel must be specified.

MolecularGraph.initcanvas! — Method

initcanvas!(canvas::Canvas, coords::AbstractArray{Float64}, boundary::Tuple)

Move and adjust the size of the molecule for drawing.