Molecular structure drawing

MolecularGraph.chargesign — Method

chargesign(charge::Int) -> String

Get a charge sign.

MolecularGraph.draw2d! — Method

draw2d!(canvas::Canvas, mol::UndirectedGraph; kwargs...)

Draw molecular image to the canvas.

MolecularGraph.is_atom_visible — Method

isatomvisible(mol::SimpleMolGraph; setting=DRAW_SETTING) -> Vector{Bool}

Return whether the atom is visible in the 2D drawing.

MolecularGraph.drawsvg — Method

drawsvg(mol::SimpleMolGraph, width::Int, height::Int)

Generate molecular structure image as a SVG format string.

width and height specifies the size of the image (width and height attribute of svg tag).

MolecularGraph.html_fixed_size — Method

html_fixed_size(svg, width, height)

Generate fixed-size HTML wrapper for the SVG element.

MolecularGraph.html_grid — Method

html_grid(svgs, cols, rowheight)

Generate grid layout HTML wrapper for the SVG elements.

The number of columns and grid row height in pixel must be specified.

MolecularGraph.initcanvas! — Method

initcanvas!(canvas::Canvas, coords::AbstractArray{Float64}, boundary::Tuple)

Move and adjust the size of the molecule for drawing.