Preprocessing

MolecularGraph.addhydrogens! — Method

addhydrogens!(mol::GraphMol) -> GraphMol

Add hydrogen nodes to the molecular graph explicitly.

Note that this function edits Atom and Bond object fields directly (see MolecularGraphModel.clone and MolecularGraphModel.clearcache!.

MolecularGraph.allhydrogens — Method

allhydrogens(mol::GraphMol) -> Set{Int}

Return a set of hydrogen nodes.

MolecularGraph.depolarize! — Method

depolarize!(mol::GraphMol)

Depolarize oxo groups except in the case that polarization is required for aromaticity.

Note that this function edits Atom and Bond object fields directly (see MolecularGraphModel.clone and MolecularGraphModel.clearcache!.

MolecularGraph.largestcomponentgraph — Method

largestcomponentgraph(mol::GraphMol) -> SubgraphView

Return largest connected component view of the molecular graph.

MolecularGraph.largestcomponentnodes — Method

largestcomponentnodes(mol::GraphMol) -> Set{Int}

Return a set of nodes in the largest connected component.

MolecularGraph.neutralizeacids! — Method

neutralizeacids!(mol::GraphMol)

Neutralize oxo(thio) acids.

Note that this function edits Atom object fields directly (see MolecularGraphModel.clone and MolecularGraphModel.clearcache!.

MolecularGraph.neutralizeoniums! — Method

neutralizeoniums!(mol::GraphMol)

Neutralize 1-3° oniums. Permanently charged quart-oniums will not be neutralized.

Note that this function edits Atom object fields directly (see MolecularGraphModel.clone and MolecularGraphModel.clearcache!.

MolecularGraph.preprocess! — Method

preprocess!(mol::GraphMol)

Default molecular preprocessing method.

Ionized acids and oniums will be neutralized (neutralizeacids! and neutralizeoniums!).

If there is several possible resonance structures, the one which have less total charge (if they are even, the one which have less maximum bond order) will be selected (depolarize! and toallenelike!).

For further preprocessing, following removehydrogens, removestereohydrogens and largestcomponentgraph would work well.

Note that this function edits Atom and Bond object fields directly (see MolecularGraphModel.clone and MolecularGraphModel.clearcache!).

MolecularGraph.removehydrogens — Method

removehydrogens(mol::GraphMol) -> GraphMol

Return molecule whose hydrogen nodes are removed.

If option all is set to false, only trivial hydrogens are removed (see trivialhydrogens). If you want to keep stereochemistry, remove hydrogens with all=false, and do removestereohydrogens.

If

MolecularGraph.toallenelike! — Method

toallenelike!(mol::GraphMol)

Convert triple bonds with charges into allene-like structure (ex. [C-][N+]#N -> C=[N+]=[N-]).

Note that this function edits Atom and Bond object fields directly (see MolecularGraphModel.clone and MolecularGraphModel.clearcache!.

MolecularGraph.trivialhydrogens — Method

trivialhydrogens(mol::GraphMol) -> Set{Int}

Return a set of trivial hydrogen nodes (light hydrogens which are uncharged, non-radical, non-stereospecific and attached to organic heavy atoms)

Stereochemistry

MolecularGraph.addstereohydrogens — Method

addstereohydrogens(mol::SMILES) -> SMILES

Return new molecule with explicit hydrogen nodes attached to stereocenters.

MolecularGraph.optimizewedges! — Method

optimizewedges!(mol::SDFile)

Optimize dashes and wedges representations. Typical stereocenters can be drawn as four bonds including only a wedge and/or a dash, so if there are too many dashes and wedges, some of them will be converted to normal single bond without changing any stereochemistry.

MolecularGraph.removestereohydrogens — Method

removestereohydrogens(mol::SMILES) -> SMILES

Return new molecule without explicit hydrogen nodes attached to the stereocenters.

MolecularGraph.setdiastereo! — Method

setdiastereo!(mol::Union{SMILES,SDFile}) -> nothing

Set diastereomerism flags to Bond.stereo fields of double bonds.

:unspecified, :cis or :trans will be assigned according to configurations of bonds labeled by lower indices. (ex. SMILES C/C=C(C)/C have 4 explicit bonds 1:C/C, 2:C=C, 3:C(C) and 4:C/C. 1st bond is prior to the other implicit hydrogen attached to 2nd atom, and 3rd bond is prior to 4th bond in index label order. 1st and 3th atom are in $cis$ position, so Bond.stereo of 2nd bond will be set to :cis.)

MolecularGraph.setstereocenter! — Method

setstereocenter!(mol::SDFile)

Set stereocenter information to Atom.stereo (:unspecified, :clockwise, :anticlockwise or :atypical). Clockwise/anticlockwise means the configuration of 2-4th nodes in index label order when we see the chiral center from the node labeled by the lowest index. If there is an implicit hydrogen, its index label will be regarded as the same as the stereocenter atom.

Note that setstereocenter! will optimize dashes and wedges by using optimizewedges! inside it. If you do not want to change the bond notation, clone the molecule before running setstereocenter!.