Preprocessing

Preprocessing

MolecularGraph.allhydrogens — Method.

allhydrogens(mol::GraphMol) -> Set{Int}

Return a set of hydrogen nodes.

source

MolecularGraph.canonicalize! — Method.

canonicalize!(mol::GraphMol)

Canonicalize molecule notation and apply the changes to the molecular property vector.

Neutralize oxo acid, 1-3° ammonium and polarized carbonyls except in the case that polarization is required for aromaticity.

Canonicalize anions next to triple bonds (ex. [C-][N+]#N -> C=[N+]=[N-])

source

MolecularGraph.depolarize! — Method.

depolarize!(mol::GraphMol)

Depolarize oxo groups except in the case that polarization is required for aromaticity.

Note that this function edits Atom object fields directly. The molecular property vector needs recalculation to apply the changes. see canonicalize!.

source

MolecularGraph.largestcomponentgraph — Method.

largestcomponentgraph(mol::GraphMol) -> GraphMol

Return largest connected component of the molecular graph.

source

MolecularGraph.largestcomponentnodes — Method.

largestcomponentnodes(mol::GraphMol) -> Set{Int}

Return a set of nodes in the largest connected component.

source

MolecularGraph.makehydrogensexplicit — Method.

makehydrogensexplicit(mol::GraphMol) -> GraphMol

Return molecule whose hydrogens are fully attached. If option all is set to false, only trivial hydrogens are removed (see trivialhydrogens).

source

MolecularGraph.makehydrogensimplicit — Method.

makehydrogensimplicit(mol::GraphMol) -> GraphMol

Return molecule whose hydrogen nodes are removed. If option all is set to false, only trivial hydrogens are removed (see trivialhydrogens).

source

MolecularGraph.neutralizeacids! — Method.

neutralizeacids!(mol::GraphMol)

Neutralize oxo(thio) acids.

Note that this function edits Atom object fields directly. The molecular property vector needs recalculation to apply the changes. see canonicalize!.

source

MolecularGraph.neutralizeoniums! — Method.

neutralizeoniums!(mol::GraphMol)

Neutralize 1-3° oniums. Permanently charged quart-oniums are not neutralized.

Note that this function edits Atom object fields directly. The molecular property vector needs recalculation to apply the changes. see canonicalize!.

source

MolecularGraph.triplebondanion! — Method.

triplebondanion!(mol::GraphMol)

Canonicalize anions next to triple bonds (ex. [C-][N+]#N -> C=[N+]=[N-]).

Note that this function edits Atom object fields directly. The molecular property vector needs recalculation to apply the changes. see canonicalize!.

source

MolecularGraph.trivialhydrogens — Method.

trivialhydrogens(mol::GraphMol) -> Set{Int}

Return a set of trivial hydrogen nodes (light hydrogens which are uncharged, non-radical, non-stereospecific and attached to organic heavy atoms)

source