Preprocessing

MolecularGraph.add_hydrogens! — Method

addhydrogens(mol::GraphMol) -> GraphMol

Return the molecule with all hydrogen nodes explicitly attached.

MolecularGraph.all_hydrogens — Method

allhydrogens(mol::SimpleMolGraph) -> Set{Int}

Return a set of all hydrogen nodes.

MolecularGraph.depolarize — Method

depolarize(mol::SimpleMolGraph; negative=:O, positive=[:C, :P]) -> Nothing

Depolarize dipole double bonds of the molecule.

MolecularGraph.deprotonate_oniums — Method

deprotonate_oniums(mol::SimpleMolGraph) -> Vector{Int}

Deprotonate onium groups of the molecule.

MolecularGraph.extract_largest_component! — Method

extract_largest_component!(mol::SimpleMolGraph) -> Nothing

Return the largest connected component of the molecular graph.

This should be useful when you want to remove salt and water molecules from the molecular graph simply. On the other hand, this can remove important components from the mixture so carefully apply this preprocess method.

MolecularGraph.kekulize — Method

kekulize(mol::SimpleMolGraph) -> Vector{Int}

Kekulize the molecule that has SMILES aromatic bonds.

SMILES allows aromatic atoms in small letters - b, c, n, o, p, s, [as] and [se]. Once these are stored in SmilesAtom.isaromatic field, then kekulize will place double bonds to satisfy valences.

Kekulization is necessary for molecules parsed from SMILES. If not kekulized, some bond valence and implicit hydrogen properties would be wrong.

MolecularGraph.largest_component_nodes — Method

largest_component_nodes(mol::SimpleMolGraph) -> Vector{Int}

Return a set of nodes in the largest connected component.

MolecularGraph.polarize — Method

polarize(mol::SimpleMolGraph; negative=:O, positive=[:N, :S]) -> Nothing

Polarize dipole double bonds of the molecule.

MolecularGraph.protonate_acids — Method

protonate_acids(mol::SimpleMolGraph) -> Vector{Int}

Protonate oxo(thio) anion groups of the molecule.

MolecularGraph.removable_hydrogens — Method

removable_hydrogens(mol::SimpleMolGraph) -> Set{Int}

Return a vector of removable hydrogen nodes.

removable means not charged, no unpaired electron, no specific mass, non-stereospecific and attached to organic heavy atoms.

MolecularGraph.remove_hydrogens! — Method

removehydrogens(mol::MolGraph) -> GraphMol

Return the molecule with hydrogen nodes removed.

If option all is set to true (default), all hydrogens will be removed, otherwise only trivial hydrogens will be removed (see trivialhydrogens).

MolecularGraph.to_allene_like — Method

to_allene_like(mol::SimpleMolGraph) -> Nothing

Standardize the molecule so that all 1,3-dipole groups are represented as allene-like structure (e.g. Diazo group [C-][N+]#N -> C=[N+]=[N-]).

MolecularGraph.to_triple_bond — Method

to_triple_bond(mol::SimpleMolGraph) -> Nothing

Standardize the molecule so that all 1,3-dipole groups are represented as triple bond and single bond (e.g. Diazo group C=[N+]=[N-] -> [C-][N+]#N).

Stereochemistry

MolecularGraph.remove_stereo_hydrogen! — Method

remove_stereo_hydrogen!(mol::SimpleMolGraph, v::Integer) -> Bool

Safely remove explicit hydrogens connected to stereocenter node v.

MolecularGraph.set_stereobond! — Method

set_stereocenter!(mol::SimpleMolGraph, bond, v1, v2, is_cis) -> Nothing

Set stereocenter information to graph properties.

MolecularGraph.set_stereocenter! — Method

set_stereocenter!(mol::SimpleMolGraph, center, looking_from, v1, v2, is_clockwise) -> Nothing

Set stereocenter information to graph properties.

MolecularGraph.stereobond_from_sdf2d! — Method

stereobond_from_sdf2d!(mol::MolGraph) -> Nothing

Set cis-trans diastereomerism information obtained from 2D SDFile.

MolecularGraph.stereobond_from_smiles! — Method

stereobond_from_smiles!(mol::MolGraph) -> Nothing

Set cis-trans diastereomerism information obtained from SMILES.

MolecularGraph.stereocenter_from_sdf2d! — Method

stereocenter_from_sdf2d!(mol::MolGraph) -> Nothing

Set stereocenter information obtained from 2D SDFile.

MolecularGraph.stereocenter_from_smiles! — Method

stereocenter_from_smiles!(mol::MolGraph) -> Nothing

Set stereocenter information obtained from SMILES.