Molecular properties

MolecularGraph.atomcounter — Method

atomcounter(mol::GraphMol) -> Dict{Symbol,Int}

Count the number of atoms and return symbol => count dict.

MolecularGraph.atomsymbol — Method

atomsymbol(mol::GraphMol) -> Vector{Symbol}

Return a vector of size $n$ representing atom symbols of 1 to $n$th atoms of the given molecule.

MolecularGraph.bondorder — Method

bondorder(mol::GraphMol) -> Vector{Int}

Return a vector of size $n$ representing bond order of 1 to $n$th bonds of the given molecule.

MolecularGraph.charge — Method

charge(mol::GraphMol) -> Vector{Int}

Return a vector of size $n$ representing atom charges of 1 to $n$th atoms of the given molecule.

MolecularGraph.connectivity — Method

connectivity(mol::GraphMol) -> Vector{Int}

Return a vector of size $n$ representing the number of total atoms (implicit and explicit) connected to 1 to $n$th atoms of the given molecule.

This property corresponds to SMARTS X query.

MolecularGraph.empiricalformula — Method

empiricalformula(mol::GraphMol) -> String

Return the empirical formula in Hill system.

MolecularGraph.explicithconnected — Method

explicithconnected(mol::GraphMol) -> Vector{Int}

Return a vector of size $n$ representing the number of explicit hydrogens connected to 1 to $n$th atoms of the given molecule.

"Explicit" means hydrogens are connected to the heavy atom as atom nodes.

MolecularGraph.fusedringmembership — Method

fusedringmembership(mol::UndirectedGraph) -> Vector{Int}

Return a vector of size $n$ representing fusedrings membership of 1 to $n$th atoms of the given molecule.

Fused ring membership is represented as a set of fused ring indices assigned to each fused rings. This means atoms that have the same fused ring index belong to the same fused ring.

MolecularGraph.fusedrings — Method

fusedrings(mol::UndirectedGraph) -> Vector{Set{Int}}

Return vectors of fused ring node sets.

A fused ring is defined as a 2-edge connected components in terms of graph theory. Spirocyclic structures are considered to be part of a fused ring.

MolecularGraph.hacceptorcount — Method

hacceptorcount(mol::GraphMol) -> Int

Return the total number of hydrogen bond acceptors (N, O and F).

MolecularGraph.hdonorcount — Method

hdonorcount(mol::GraphMol) -> Int

Return the total number of hydrogen bond donors (O and N attached to hydrogens).

MolecularGraph.heavyatomconnected — Method

heavyatomconnected(mol::GraphMol) -> Vector{Int}

Return a vector of size $n$ representing the number of non-hydrogen atoms connected to 1 to $n$th atoms of the given molecule.

MolecularGraph.heavyatomcount — Method

heavyatomcount(mol::GraphMol) -> Int

Return the total number of non-hydrogen atoms.

MolecularGraph.hybridization — Method

hybridization(mol::GraphMol) -> Vector{Int}

Returns a vector of size $n$ representing the orbital hybridization symbols (:sp3, :sp2, :sp or :none) of 1 to $n$th atoms of the given molecule.

The hybridization value in inorganic atoms and non-typical organic atoms will be :none (e.g. s, sp3d and sp3d2 orbitals).

MolecularGraph.hydrogenconnected — Method

hydrogenconnected(mol::GraphMol) -> Vector{Int}

Return a vector of size $n$ representing the number of total hydrogens (implicit and explicit) connected to 1 to $n$th atoms of the given molecule.

This property corresponds to SMARTS H query.

MolecularGraph.implicithconnected — Method

implicithconnected(mol::GraphMol) -> Vector{Int}

Return a vector of size $n$ representing the number of implicit hydrogens connected to 1 to $n$th atoms of the given molecule.

"Implicit" means hydrogens are not represented as graph nodes, but it can be infered from the intrinsic valence of typical organic atoms.

MolecularGraph.isaromatic — Method

isaromatic(mol::GraphMol) -> Vector{Bool}

Returns a vector of size $n$ representing whether 1 to $n$th atoms of the given molecule belong to an aromatic ring or not.

Some kind of aromaticity resulting from long conjugated chains and charge delocalization may be unrecognizable. Also, non-classical aromaticity such as Moebius aromaticity is not considered.

MolecularGraph.isaromaticbond — Method

isaromaticbond(mol::GraphMol) -> Vector{Bool}

Returns a vector of size $n$ representing whether 1 to $n$th bonds of the given molecule belong to an aromatic ring or not.

See isaromatic.

MolecularGraph.isringatom — Method

isringatom(mol::UndirectedGraph) -> Vector{Bool}

Return a vector of size $n$ representing whether 1 to $n$th atoms of the given molecule belong to a ring or not.

MolecularGraph.isringbond — Method

isringbond(mol::UndirectedGraph) -> Vector{Bool}

Return a vector of size $n$ representing whether 1 to $n$th bonds of the given molecule belong to a ring or not.

MolecularGraph.isrotatable — Method

isrotatable(mol::GraphMol)

Return a vector of size $n$ representing whether 1 to $n$th bonds of the given molecule are rotatable or not.

MolecularGraph.lonepair — Method

lonepair(mol::GraphMol) -> Vector{Union{Int,Nothing}}

Return a vector of size $n$ representing the number of lone pairs of 1 to $n$th atoms of the given molecule.

The number of lone pair in inorganic atoms would be nothing. The result can take negative value if the atom has empty shells (e.g. B).

MolecularGraph.molecularformula — Method

molecularformula(mol::GraphMol) -> String

Return the molecular formula in Hill system.

MolecularGraph.multiplicity — Method

multiplicity(mol::GraphMol) -> Vector{Int}

Return a vector of size $n$ representing atom multiplicities of 1 to $n$th atoms of the given molecule (1: non-radical, 2: radical, 3: biradical).

MolecularGraph.nodedegree — Method

nodedegree(mol::GraphMol) -> Vector{Int}

Return a vector of size $n$ representing the node degree of the molecular graph of 1 to $n$th atoms of the given molecule.

This property corresponds to SMARTS D query.

MolecularGraph.pielectron — Method

pielectron(mol::GraphMol) -> Vector{Int}

Returns a vector of size $n$ representing the number of $\pi$ electrons of 1 to $n$th atoms of the given molecule.

The counting of $\pi$ electrons is based on the following rules.

Any atom incident to a double bond -> +1

Any atom incident to two double bond -> +2

Any atom incident to a triple bond -> +2

Any other uncharged N, O or S that are neighbor of multiple bonds -> +2

These rules are applied for only typical organic atoms. The values for inorganic atoms will be 0.

MolecularGraph.precalculate! — Method

precalculate!(mol::GraphMol)

Convenient method to pre-calculate and cache performance bottleneck descriptors.

MolecularGraph.rotatablecount — Method

rotatablecount(mol::GraphMol) -> Int

Return the total number of rotatable bonds.

MolecularGraph.smallestsssr — Method

smallestsssr(mol::UndirectedGraph) -> Vector{Int}

Return a vector of size $n$ representing the size of the smallest sssr that 1 to $n$th atoms of the given molecule belong to.

This property corresponds to SMARTS r query.

MolecularGraph.sssr — Function

sssr(mol::GraphMol) -> Vector{Vector{Int}}

Return vectors of ring nodes representing small set of smallest rings (SSSR).

See Graph.minimumcyclebasis.

MolecularGraph.sssrbondmembership — Method

sssrbondmembership(mol::GraphMol) -> Vector{Set{Int}}

Return a vector of size $n$ representing sssr membership of 1 to $n$th bonds of the given molecule.

SSSR membership is represented as a set of SSSR indices assigned to each rings. This means bonds that have the same SSSR index belong to the same SSSR.

MolecularGraph.sssrcount — Method

sssrcount(mol::UndirectedGraph) -> Vector{Int}

Return a vector of size $n$ representing the number of sssr that 1 to $n$th atoms of the given molecule belong to.

This property corresponds to SMARTS R query.

MolecularGraph.sssrmembership — Method

sssrmembership(mol::GraphMol) -> Vector{Set{Int}}

Return a vector of size $n$ representing sssr membership of 1 to $n$th atoms of the given molecule.

SSSR membership is represented as a set of SSSR indices assigned to each rings. This means atoms that have the same SSSR index belong to the same SSSR.

MolecularGraph.valence — Method

valence(mol::GraphMol) -> Vector{Union{Int,Nothing}}

Return a vector of size $n$ representing the intrinsic valence of 1 to $n$th atoms of the given molecule.

The number of implicit hydrogens would be calculated based on the valence. The valence value in inorganic atoms would be nothing. This property corresponds to SMARTS v query.

MolecularGraph.wclogp — Method

wclogp(mol::GraphMol) -> Float64

Return predicted logP value calculated by using Wildman and Crippen method.

Reference

Wildman, S. A. and Crippen, G. M. (1999). Prediction of Physicochemical Parameters by Atomic Contributions. Journal of Chemical Information and Modeling, 39(5), 868–873. https://doi.org/10.1021/ci990307l

MolecularGraph.wclogptype — Method

wclogptype(mol::GraphMol)

Return Wildman-Crippen LogP atom types.

MolecularGraph.inchi — Method

inchi(molblock::String) -> String
inchi(mol::GraphMol) -> String

Generate InChI string from molblock string or molecule

MolecularGraph.inchikey — Method

inchikey(inchi::String) -> String
inchikey(mol::GraphMol) -> String

Generate InChI key from InChI string or molecule