Molecular properties

MolecularGraph.atomcounter — Method

atomcounter(mol::GraphMol) -> Dict{Symbol,Int}

Count the number of atoms and return symbol => count dict.

MolecularGraph.atomsymbol — Method

atomsymbol(mol::GraphMol) -> Vector{Int}

Return the size $n$ vector of atom symbols within the molecule that have $n$ atoms.

MolecularGraph.bondorder — Method

bondorder(mol::GraphMol) -> Vector{Int}

Return the size $n$ vector of bond orders within the molecule that have $n$ bonds.

MolecularGraph.charge — Method

charge(mol::GraphMol) -> Vector{Int}

Return the size $n$ vector of atom charges within the molecule that have $n$ atoms.

MolecularGraph.connectivity — Method

connectivity(mol::GraphMol) -> Vector{Int}

Return the size $n$ vector of the total number of adjacent atoms (including implicit hydrogens) within the molecule that have $n$ atoms.

connectivity values correspond to SMARTS X property.

MolecularGraph.empiricalformula — Method

empiricalformula(mol::GraphMol) -> String

Return the empirical formula in Hill system.

MolecularGraph.explicithconnected — Method

explicithconnected(mol::GraphMol) -> Vector{Int}

Return the size $n$ vector of the number of adjacent explicit hydrogen nodes within the molecule that have $n$ atoms.

MolecularGraph.hacceptorcount — Method

hacceptorcount(mol::GraphMol) -> Int

Return the number of hydrogen bond acceptors (N, O and F).

MolecularGraph.hdonorcount — Method

hdonorcount(mol::GraphMol) -> Int

Return the number of hydrogen bond donors (O and N attached to hydrogens).

MolecularGraph.heavyatomconnected — Method

heavyatomconnected(mol::GraphMol) -> Vector{Int}

Return the size $n$ vector of the number of adjacent non-hydrogen atoms within the molecule that have $n$ atoms.

MolecularGraph.heavyatomcount — Method

heavyatomcount(mol::GraphMol) -> Int

Return the total number of non-hydrogen atoms.

MolecularGraph.hybridization — Method

hybridization(mol::GraphMol) -> Vector{Int}

Return the size $n$ vector of orbital hybridization symbols (:sp3, :sp2, :sp, :none) within the molecule that have $n$ atoms.

These hybridizations are for only typical organic atoms. Inorganic atoms and other orbitals like s, sp3d and sp3d2 will be :none.

MolecularGraph.hydrogenconnected — Method

hydrogenconnected(mol::GraphMol) -> Vector{Int}

Return the size $n$ vector of the total number of hydrogens attached to the atom within the molecule that have $n$ atoms.

hydrogenconnected values correspond to SMARTS H property.

MolecularGraph.implicithconnected — Method

implicithconnected(mol::GraphMol) -> Vector{Int}

Return the size $n$ vector of the number of implicit hydrogens within the molecule that have $n$ atoms.

MolecularGraph.isaromatic — Method

isaromatic(mol::GraphMol)

Return the vector whether the atom belongs to an aromatic ring.

Note that aromaticity described here means simplified binary descriptor (aromatic or not) based on classical Huckel's rule. This is intended for use in some kind of pharmaceutical researches. Non-classical aromaticity such as Moebius aromaticity is not considered.

MolecularGraph.isrotatable — Method

isrotatable(mol::GraphMol)

Return whether the bonds are rotatable or not.

MolecularGraph.lonepair — Method

lonepair(mol::GraphMol) -> Vector{Union{Int,Nothing}}

Return the size $n$ vector of the number of lone pairs within the molecule that have $n$ atoms.

Note that implicit hydrogens are available for only organic atoms. The lonepair value of inorganic atoms would be nothing. The result can take negative value if the atom has empty valence shells.

MolecularGraph.molecularformula — Method

molecularformula(mol::GraphMol) -> String

Return the molecular formula in Hill system.

MolecularGraph.multiplicity — Method

multiplicity(mol::GraphMol) -> Vector{Int}

Return the size $n$ vector of atom multiplicities within the molecule that have $n$ atoms (1: non-radical, 2: radical, 3: biradical).

MolecularGraph.nodedegree — Method

nodedegree(mol::GraphMol) -> Vector{Int}

Return the size $n$ vector of node degrees within the molecular graph that have $n$ atom nodes.

nodedegree values correspond to SMARTS D property.

MolecularGraph.pielectron — Method

pielectron(mol::GraphMol) -> Vector{Int}

Return the size $n$ vector of the number of $\pi$ electrons (including implicit hydrogens) within the molecule that have $n$ atoms.

The $\pi$ electron enumelation is based on the following simple rules.

Any atom incident to a double bond -> +1

Any atom incident to two double bond -> +2

Any atom incident to a triple bond -> +1

Uncharged N and O not incident to any multiple bonds and adjacent to atoms described above -> +2

These rules are applied for only typical organic atoms. The values for inorganic atoms will be 0.

MolecularGraph.rotatablecount — Method

rotatablecount(mol::GraphMol) -> Int

Return the number of rotatable bonds.

MolecularGraph.valence — Method

valence(mol::GraphMol) -> Vector{Union{Int,Nothing}}

Return the size $n$ vector of intrinsic valences (with considering implicit hydrogens) within the molecule that have $n$ atoms.

valence values correspond to SMARTS v property. valence values of inorganic atoms would be nothing.