Molecular properties

MolecularGraph.countatoms — Method

countatoms(mol::GraphMol) -> Dict{Symbol,Int}

Count the number of atoms and return symbol => count dict.

MolecularGraph.empiricalformula — Method

empiricalformula(mol::GraphMol) -> String

Return the empirical formula in Hill system.

MolecularGraph.hacceptorcount — Method

hacceptorcount(mol::GraphMol) -> Int

Return the number of hydrogen bond acceptors (N, O and F).

MolecularGraph.hdonorcount — Method

hdonorcount(mol::GraphMol) -> Int

Return the number of hydrogen bond donors (O and N attached to hydrogens).

MolecularGraph.heavyatomcount — Method

heavyatomcount(mol::GraphMol) -> Vector{Int}

Return the size $n$ vector of the number of adjacent non-hydrogen atoms within the molecule that have $n$ atoms.

MolecularGraph.lonepair — Method

lonepair(mol::GraphMol) -> Vector{Union{Int,Nothing}}

Return the size $n$ vector of the number of lone pairs within the molecule that have $n$ atoms.

Note that implicit hydrogens are available for only organic atoms. The lonepair value of inorganic atoms would be nothing. The result can take negative value if the atom has empty valence shells.

MolecularGraph.molecularformula — Method

molecularformula(mol::GraphMol) -> String

Return the molecular formula in Hill system.

MolecularGraph.rotatablecount — Method

rotatablecount(mol::GraphMol) -> Int

Return the number of rotatable bonds.