Molecular properties

MolecularGraph.empiricalformula — Method

empiricalformula(mol::GraphMol) -> String

Return the empirical formula in Hill system.

MolecularGraph.hacceptorcount — Method

hacceptorcount(mol::GraphMol) -> Int

Return the number of hydrogen bond acceptors (N, O and F).

MolecularGraph.hdonorcount — Method

hdonorcount(mol::GraphMol) -> Int

Return the number of hydrogen bond donors (O and N attached to hydrogens).

MolecularGraph.heavyatomcount — Method

heavyatomcount(mol::GraphMol) -> Int

Return the total number of non-hydrogen atoms.

MolecularGraph.molecularformula — Method

molecularformula(mol::GraphMol) -> String

Return the molecular formula in Hill system.

MolecularGraph.precalculate! — Method

precalculate!(mol::GraphMol)

Convenient method to pre-calculate and cache performance bottleneck descriptors.

MolecularGraph.rotatablecount — Method

rotatablecount(mol::GraphMol) -> Int

Return the number of rotatable bonds.

MolecularGraph.sssr — Function

sssr(mol::UndirectedGraph) -> Vector{Vector{Int}}

Calculate Small set of smallest rings (SSSR).

This returns a vector of rings represented as a vector of atom node index along with the cycle path.

MolecularGraph.sssrmembership — Function

sssrmembership(mol::UndirectedGraph) -> Vector{Set{Int}}

Return size n vector of SSSR membership set where n is the number of atom nodes.

The numbers in the set correspond to the index of sssr that the node belongs.

MolecularGraph.wclogp — Method

wclogp(mol::GraphMol) -> Float64

Return predicted logP value calculated by using Wildman and Crippen method.

Reference

Wildman, S. A. and Crippen, G. M. (1999). Prediction of Physicochemical Parameters by Atomic Contributions. Journal of Chemical Information and Modeling, 39(5), 868–873. https://doi.org/10.1021/ci990307l

MolecularGraph.inchi — Method

inchi(molblock::String) -> String
inchi(mol::GraphMol) -> String

Generate InChI string from molblock string or molecule

MolecularGraph.inchikey — Method

inchikey(inchi::String) -> String
inchikey(mol::GraphMol) -> String

Generate InChI key from InChI string or molecule