Molecular queries

MolecularGraph.convertnotquery — Method

convertnotquery(mol::QueryMol) -> QueryMol

Convert :not query to :or if possible.

[!#1] would be *. Typical substructure mining tasks do not care stereochemistry,

so all hydrogen nodes can be removed.)

following tidyformula is necessary.

MolecularGraph.filter_queries — Method

filter_queries(qr::DictDiGraph, mol::GraphMol) -> DictDiGraph

Filter query relationship diagram by the given molecule. The filtered diagram represents query relationship that the molecule have.

MolecularGraph.findallformula — Method

findallformula(q::QueryFormula, key::Symbol) -> Any

Find all formula values that matched the given key regardless of logical operator roles.

MolecularGraph.findformula — Method

findformula(q::QueryFormula, key::Symbol) -> Any

Find value of the given key.

If deep=true is set, it will also search under :or sets to match multiple formulae and return the value after applying the aggregate function to them.

MolecularGraph.inferatomaromaticity! — Method

inferatomaromaticity!(qmol::QueryMol)

Infer aromaticity of atoms from connected bonds.

This directly add :isaromatic formulas to query atoms in the query molecule.

MolecularGraph.query_relationship — Method

query_relationship(;sourcefile=DEFAULT_QUERY_RELATIONS) -> DictDiGraph

Generate query relationship diagram.

MolecularGraph.querymol — Method

querymol(mol::SubgraphView{QueryMol}) -> QueryMol

Generate a new QueryMol from a substructure view.

Graph property caches and attributes are not inherited.

MolecularGraph.querymol — Method

querymol(atoms::Vector{Atom}, bonds::Vector{Bond}) -> GraphMol

Generate QueryMol that has the given atom objects and edge objects.

MolecularGraph.querymol — Method

querymol() -> QueryMol

Generate empty QueryMol.

MolecularGraph.removehydrogens — Method

removehydrogens(mol::QueryMol) -> QueryMol

Return the molecular query with hydrogen nodes removed.

Hydrogen nodes would be represented as [*Hx] in the connected heavy atoms.

(e.g. CN[H] -> C[N;H1,H2,H3,H4])

MolecularGraph.tidyformula — Method

tidyformula(fml::QueryFormula) -> QueryFormula

Return tidy formulae.

associative formulae will be juxtaposed (ex. :and => (A, :and => (B, C)) -> :and => (A, B, C))

distributive formulae will be factored out (ex. :or => (:and => (A, B), :and => (A, C)) -> :and => (A, :or => (B, C)))

Absorption (ex. :and => (A, :or => (A, B)) -> A

:any absorbs everything (ex. :and => (:any => true, A) -> A, :or => (:any => true, A) -> :any => true

:not would be inverted if possible (ex. :not => (:A => true) -> :A => false)

:not has the highest precedence in SMARTS, but only in the case like [!C], De Morgan's law will be applied to remove :and under :not. (ex. :not => (:and => (:atomsymbol => :C, :isaromatic => false) -> :or => (:not => (:atomsymbol => :C), isaromatic => true)