Stereochemistry

MolecularGraph.safe_stereo_hydrogen! — Method

safe_stereo_hydrogen!(mol::SimpleMolGraph, v::Integer) -> Bool

Rearrange stereocenter properties to safely remove stereo hydrogen nodes and return the hydrogen node index.

This function is called inside rem_vertex! and rem_vertices! functions to safely remove hydrogen nodes while preserving stereocenter information.

MolecularGraph.set_stereobond! — Method

set_stereocenter!(mol::SimpleMolGraph, bond, v1, v2, is_cis) -> Nothing

Set stereocenter information to graph properties.

MolecularGraph.set_stereocenter! — Method

set_stereocenter!(mol::SimpleMolGraph, center, looking_from, v1, v2, is_clockwise) -> Nothing

Set stereocenter information to graph properties.

MolecularGraph.stereobond_from_sdf2d! — Method

stereobond_from_sdf2d!(mol::MolGraph) -> Nothing

Set cis-trans diastereomerism information obtained from 2D SDFile.

MolecularGraph.stereobond_from_smiles! — Method

stereobond_from_smiles!(mol::MolGraph) -> Nothing

Set cis-trans diastereomerism information obtained from SMILES.

MolecularGraph.stereocenter_from_sdf2d! — Method

stereocenter_from_sdf2d!(mol::MolGraph) -> Nothing

Set stereocenter information obtained from 2D SDFile.

MolecularGraph.stereocenter_from_smiles! — Method

stereocenter_from_smiles!(mol::MolGraph) -> Nothing

Set stereocenter information obtained from SMILES.